+++ files/gitlab-cookbooks/package/libraries/config/roles/application.rb
@@ -1,4 +1,3 @@
-#
 # Copyright:: Copyright (c) 2017 GitLab Inc.
 # License:: Apache License, Version 2.0
 #
@@ -15,9 +14,11 @@
 # limitations under the License.
 #
 
-require 'uri'
-require 'digest'
-require_relative 'helpers/redhat_helper'
-require_relative 'helpers/secrets_helper'
-require_relative 'helpers/version_helper'
-require_relative 'helpers/quote_helper'
+module ApplicationRole
+  def self.load_role
+    return unless Gitlab['application_role']['enable']
+
+    Gitlab['gitlab_rails']['enable'] = true
+    Services.enable_group('rails')
+  end
+end