+++ files/gitlab-cookbooks/package/libraries/config/roles/geo_primary.rb
@@ -1,4 +1,3 @@
-#
 # Copyright:: Copyright (c) 2017 GitLab Inc.
 # License:: Apache License, Version 2.0
 #
@@ -15,9 +14,16 @@
 # limitations under the License.
 #
 
-require 'uri'
-require 'digest'
-require_relative 'helpers/redhat_helper'
-require_relative 'helpers/secrets_helper'
-require_relative 'helpers/version_helper'
-require_relative 'helpers/quote_helper'
+module GeoPrimaryRole
+  def self.load_role
+    return unless Gitlab['geo_primary_role']['enable']
+
+    Gitlab['gitlab_rails']['geo_primary_role_enabled'] = true
+    Gitlab['postgresql']['sql_replication_user'] ||= 'gitlab_replicator'
+    Gitlab['postgresql']['wal_level'] = 'hot_standby'
+    Gitlab['postgresql']['max_wal_senders'] ||= 10
+    Gitlab['postgresql']['wal_keep_segments'] ||= 10
+    Gitlab['postgresql']['max_replication_slots'] ||= 1
+    Gitlab['postgresql']['hot_standby'] = 'on'
+  end
+end