+++ spec/chef/recipes/config_spec.rb
@@ -10,7 +10,6 @@ describe 'gitlab::config' do
 
   shared_examples 'regular services are disabled' do
     it 'disables regular services' do
-      expect(node['gitlab']['gitlab-rails']['enable']).to eq false
       expect(node['gitlab']['unicorn']['enable']).to eq false
       expect(node['gitlab']['sidekiq']['enable']).to eq false
       expect(node['gitlab']['gitlab-workhorse']['enable']).to eq false
@@ -74,6 +73,10 @@ describe 'gitlab::config' do
         stub_gitlab_rb(
           redis_sentinel_role: {
             enable: true
+          },
+          redis: {
+            master_password: 'PASSWORD',
+            master_ip: '10.0.0.0'
           }
         )
       end
@@ -83,7 +86,7 @@ describe 'gitlab::config' do
       it 'only sentinel is enabled' do
         expect(node['gitlab']['sentinel']['enable']).to eq true
         expect(node['gitlab']['redis']['enable']).to eq false
-        expect(node['gitlab']['redis-exporter']['enable']).to eq true
+        expect(node['gitlab']['redis-exporter']['enable']).to eq false
         expect(node['gitlab']['node-exporter']['enable']).to eq true
         expect(node['gitlab']['logrotate']['enable']).to eq true
       end