+++ CHANGELOG.md
@@ -5,14 +5,21 @@ omnibus-gitlab repository.
 
 10.0.0
 
+- Add option to override the hostname for remote syslog
 - Add backup_timeout argument to geo db replication command
 - Remove sensitive params from the NGINX access logs 6983fe59
+- Upgrade rubygems to 2.6.13 4650cd70a
 - Add option to pass EXTERNAL_URL during installation d0f30ef2
   * Saves users from manually editing gitlab.rb just to set the URL and hence
     makes installation process easier
+- Remove TLSv1 from the list of accepted ssl protocols
 - Moved the settings handling into the package cookbook and reduced code duplication in settings
+- Move the GitLab HA roles into their own files, and switch default services to be enabled by a Default role
 - Remove geo_bulk_notify_worker_cron 44def4b5
 - Remove support for the old git_data_dir configuration option
+- Rework `single_quote` helper as `quote` that can handle escaping
+  strings with embedded quotes fdc6a93- Add omniauth sync_profile_from_provider and sync_profile_attributes configuration
+- Add omniauth sync_profile_from_provider and sync_profile_attributes configuration
 
 9.5.5
 - Add more options to repmgr.conf template